requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
attrs==25.1.0
beautifulsoup4==4.13.3
binaryornot==0.4.4
blinker==1.9.0
bokeh==3.6.3
Brotli==1.1.0
cachetools==5.5.2
cerebras_cloud_sdk==1.28.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
compress-pickle==2.1.0
contourpy==1.3.1
cookiecutter==2.6.0
cssselect2==0.8.0
cycler==0.12.1
distro==1.9.0
entrypoints==0.4
exceptiongroup==1.2.2
Faker==37.1.0
favicon==0.7.0
fonttools==4.57.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
htbuilder==0.9.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lxml==5.3.2
Markdown==3.8
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
narwhals==1.29.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==6.0.1
prometheus_client==0.21.1
protobuf==5.29.3
pyarrow==19.0.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pydyf==0.11.0
Pygments==2.19.1
pymdown-extensions==10.14.3
pyparsing==3.2.3
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
st-annotated-text==4.0.2
st-theme==1.2.3
streamlit==1.43.2
streamlit-antd-components==0.3.2
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-community-navigation-bar==4.6.1
streamlit-embedcode==0.1.2
streamlit-extras==0.6.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.3.0
streamlit-option-menu==0.4.0
streamlit-plugins==0.5.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==9.0.0
text-unidecode==1.3
tinycss2==1.4.0
tinyhtml5==2.0.0
toml==0.10.2
tornado==6.4.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
validators==0.34.0
watchdog==6.0.0
weasyprint==65.1
webencodings==0.5.1
xyzservices==2025.1.0
zopfli==0.2.3.post1